Make TerritoryUpdateThread an inner class of TerritoryManager
