+++ common/src/main/java/com/wynntils/core/webapi/TerritoryManager.java
@@ -7,6 +7,7 @@ package com.wynntils.core.webapi;
 import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
+import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.CoreManager;
 import com.wynntils.core.webapi.request.RequestBuilder;
 import com.wynntils.core.webapi.request.RequestHandler;
@@ -100,5 +101,32 @@ public class TerritoryManager extends CoreManager {
 
     public static Set<TerritoryPoi> getTerritoryPois() {
         return territoryPois;
+    }
+
+    private static class TerritoryUpdateThread extends Thread {
+        private static final int TERRITORY_UPDATE_MS = 15000;
+
+        public TerritoryUpdateThread(String name) {
+            super(name);
+        }
+
+        @Override
+        public void run() {
+            RequestHandler handler = new RequestHandler();
+
+            try {
+                Thread.sleep(TERRITORY_UPDATE_MS);
+                while (!isInterrupted()) {
+                    tryLoadTerritories(handler);
+                    handler.dispatch();
+
+                    // TODO: Add events
+                    Thread.sleep(TERRITORY_UPDATE_MS);
+                }
+            } catch (InterruptedException ignored) {
+            }
+
+            WynntilsMod.info("Terminating territory update thread.");
+        }
     }
 }
